Reorganize install order so that plugin be installed last
Rationale: in development mode, make install stops with an error when
trying to install the libnautilus-actions extension library. This is
normal as I develop with my own user, not root. As the glade file has
to be installed in its configured location, it is easyer to test the
UI if it is installed before the error occurs.