Definitively fix xds_atom_value() return value
After validation both on 32 and 64 bit systems, it appears that the return buffer
almost always contains trailing trash characters. We so have to take the actual
length value to put a '\0' character at the end of the string.
Do not too understand the documentation comment which talks about 4bytes chunks...