Do some more tests on virtual methods handling
Rationale: The problem to solve is how to make sure all virtual methods of a class hierarchy
will be actually invoked.

It happends that this depends of whether an intemediate class which would not implement a
virtual method set this virtual to NULL or not in its class initialization.

If the virtual method pointer is set to NULL, then GObject library recognizes that the
method is not implemented, and this actually breaks the 'call parent method' paradigm.

Contrarily, if the virtual method is not set to NULL, then the GObject library automatically
jumps to the next parent which has implemented the method. In other words, using the 'call
parent method' paradigm implies that we never set a not-used method pointer to NULL.